batch: Move line mapping type
match.py owned both the matching algorithm and the reusable mapping value
object. Many modules imported LineMapping from the algorithm module only
for annotations or stored context.

The new line_mapping module owns the mapping protocol and lifecycle.
Matching code now imports that type privately while callers import the data
boundary directly.

This commit separates the mapping type from the matching algorithm.

The match module now focuses on occurrence scanning, LIS anchor selection,
and public matching entry points.
