+++ src/git_stage_batch/batch/baseline_edits.py
@@ -11,7 +11,7 @@ from ..core.line_selection import LineRanges, LineSelection, coerce_line_ranges
 from ..exceptions import MergeError as _MergeError
 from ..i18n import _
 from ..utils.text import normalize_line_endings
-from .match import LineMapping
+from .line_mapping import LineMapping
 from .merge_candidates import MergeResolution as _MergeResolution
 
 if TYPE_CHECKING:
