+++ src/git_stage_batch/batch/discard.py
@@ -11,7 +11,8 @@ from .baseline_correspondence import (
 from .discard_reversal import (
     reverse_presence_constraints as _reverse_batch_presence_constraints,
 )
-from .match import LineMapping, match_lines
+from .line_mapping import LineMapping
+from .match import match_lines
 from .realized_entries import (
     RealizedEntry as _RealizedEntry,
     _RealizedEntries,
